Change how the message is handled, so it all lives in the index-file
This change is made to make sure everything is rendered in the index-file and pre-exists before the actual HTML is loaded.
Before there was a functioncall in the HTML, so then when the HTML is loaded, it has to go back to the index file to execute the function,
To then go back to the HTML with the function output... This is not a good idea, so with this change (put all the information available inside
the -variable) all the info already exists when the HTML is loaded, so it can just use the information to put it there.